Ignore lint errors on CI elisp files
In Emacs < 26.1, checkdoc needs a workaround to call
checkdoc-create-error-function, and it is impossible to avoid
the prefix error of package-lint if you define the variable in
checkdoc-runner.el.

Since they will be only used in batch files and not usually
loaded in interactive Emacs sessions, package-lint errors
won't matter, so I will ignore them.