Add a command to install pre-commit hook
This commit also contains restructuring of the code.
It restructures the directories and default.nix.

Now you don't have to add '-A main' when you install the program.
It installs the elinter executable without specifying an attribute.
It still supports installing the file linter.

The default linter set is moved from emacs.nix to default.nix.