Set shell options on errors
In most situations, `set -euo pipefail` provides an expected behaviour
on errors.

In the byte-compile script, variables such as PACKAGE_ELISP_FILES
must be set, so `-u` is necessary.

Some scripts in this project, however, don't want immediate error exit
to happen, e.g. for counting the number of errors, so the `-e` option
is not always enabled.