feat!: Extend each script with a Emacs version matrix for CI
For linting, it is better to use an Emacs build as new as
possible, so emacs-snapshot would be suitable.

For byte-compiling and testing, it is often important to test with older Emacs
versions supported by the package. Now each script has `matrix` attribute,
which provides variants built with supported Emacs versions.
These cannot be exposed to the flake as packages, because the version list
depends on an IFD.

Also added `compile` option to script. This forces twist to compile local packages.
Now each script definition is an attrset, not a string.